Refactor code in cms example site
