+++ cms-examples/client/app/app.module.ts
@@ -1,15 +1,16 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { HttpModule } from '@angular/http';
 import { NgModule, CUSTOM_ELEMENTS_SCHEMA, ComponentFactoryResolver } from '@angular/core';
+import { registerPageType, CoreModule } from '@angular-cms/core';
 
-import { AppComponent } from './app.component';
 import { PagesModule } from './pages/pages.module';
 import * as register from './pages/register';
 
+import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app.routing';
-import { TestInjectService, BlogTypeSelectionFactory } from './pages/blog/test.service';
-import { registerPageType, CoreModule } from '@angular-cms/core';
-import { LayoutComponent } from './pages/shared/layout/layout.component';
+
+import { LayoutComponent } from './shared/layout/layout.component';
+import { BlogTypeSelectionFactory } from './pages/blog/blog-type-selection.factory';
 
 registerPageType(register);
 
@@ -26,7 +27,6 @@ registerPageType(register);
     LayoutComponent
   ],
   providers:[
-    TestInjectService,
     BlogTypeSelectionFactory
   ],
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
