+++ cms-examples/client/app/app.routing.ts
@@ -1,10 +1,8 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-
-import { LayoutComponent } from './pages/shared/layout/layout.component';
+import { LayoutComponent } from './shared/layout/layout.component';
 
 const appRoutes: Routes = [
-  // { path: '', component: HomeComponent },
   {
     path: 'cms',
     loadChildren: '@angular-cms/portal/src/cms.module#CmsModule',
