--- cms-examples/client/app/pages/shared/layout/layout.component.scss
@@ -1,1 +1,0 @@
-@import '../../../../assets/css/style.css';
