+++ cms-examples/client/app/shared/layout/layout.component.html
