+++ cms-examples/client/app/shared/layout/layout.component.scss
@@ -1,0 +1,1 @@
+@import '../../../assets/css/style.css';
